Correct bug in calculation of record length
There was a mistake in the usage of vsize and record length. It turns
out that vsize for a record variable is defined as the "record size",
which actually is the sum of the length of *all* record variables for
a given record. Then the length of a record is defined as the vsize plus
the offset. This mistake in the calculations caused an error in the
content-length header that gets set out.

This patch should fix the length calculations.